Damar

+++ app/Http/Controllers/ProductController.php
@@ -19,7 +19,7 @@ class ProductController extends Controller
         return view('product.index', [
             'title' => 'Product', 
             'products'=> Product::all()
-          ]);
+        ]);
     }
 
     /**
@@ -47,6 +47,7 @@ class ProductController extends Controller
      */
     public function store(Request $request)
     {
+        return response()->json('ms');
         $product_data = $request->validate([
             'name'=>['required'],
             'price'=>['required'],
@@ -54,7 +55,10 @@ class ProductController extends Controller
             'image'=>['file' , 'max: 1000'],
             'description'=>['required'],
         ]);
-        $product_data['image']=$request->file('image')->store('product_image');
+
+        if ($request->file('image')) {
+            $product_data['image']=$request->file('image')->store('product_image');
+        }
         Product::create($product_data);
         return redirect('/product'); 
     }
@@ -108,10 +112,10 @@ class ProductController extends Controller
             'image'=>['file' , 'max: 1000'],
             'description'=>['required'],
         ]);
-        if($request->old_image) {
+        if ($request->old_image && $request->file('image')) {
             Storage::delete("/$request->old_image");
+            $product_data['image']=$request->file('image')->store('product_image');
         }
-        $product_data['image']=$request->file('image')->store('product_image');
         Product::where('id', $id)->update($product_data);
         return redirect('/product'); 
     }
